Add event loop
and so we can receive error events, and don't need the *_checked calls.
actually understanding the error events remains to be implemented
though.

also, as a result, i also don't need the c23 bs, and can go right back
to c90. i also took the time to refactor some of those out parameters
into return values, only connect() really needed them. i could factor it
out there too but i think this is fine.

i know that Xt has XtAppAddTimeOut, and for a bit i thought this was an
X11 feature that i could use from xcb. sadly, as far as i can tell, this
isn't the case, rather it's implemented in Xt's own event processing,
which i'm not using here. so i instead combined xcb_poll_for_event()
with a poll() timeout and it seems to work perfectly for my purposes. it
might need extra handling if i ever enable more events, but for now this
is fine.