Add ext-session-lock-manager-v1 binding
and a whoooole lot of extra supporting code! ext-session-lock-manager-v1
is part of the ext-session-lock-v1 extension to the wayland protocol,
defined in the wayland-protocols package, but that package only provides
the xml descriptions, from which a project is intended to generate the
header and object and link it in. i've decided to make a shared library
out of it, for fun. i've also decided to build it with meson and put all
that stuff under the mit license like wayland-protocols itself because i
decided that felt right. also, meson faq says if i want to glob i need
to shell out, so i made a shell script and added a shellcheck hook and
oops apparently .envrc counts as a shell script that matters. it seems
direnv always runs it as a bash script so i added a bash shebang to make
shellcheck happy. i also decided to refactor the nix code a bunch to tie
the derivations together more easily.